+++ denops/@ddc-sources/skkeleton_okuri/main.ts
@@ -1,13 +1,13 @@
-import { getOkuriStr } from "../skkeleton/okuri.ts";
-import { okuriSplits } from "./skkeleton/okurisplits.ts";
+import { getOkuriStr } from "../../skkeleton/okuri.ts";
+import { okuriSplits } from "./okurisplits.ts";
 
 import {
   BaseSource,
   type GatherArguments,
   type GetCompletePositionArguments,
   type OnCompleteDoneArguments,
-} from "jsr:@shougo/ddc-vim@~9.1.0/source";
-import { type Item } from "jsr:@shougo/ddc-vim@~9.1.0/types";
+} from "@shougo/ddc-vim/source";
+import { type Item } from "@shougo/ddc-vim/types";
 
 type Never = Record<PropertyKey, never>;
