+++ denops/@ddc-sources/skkeleton_okuri/okurisplits_test.ts
@@ -1,6 +1,6 @@
 import { okuriSplits } from "./okurisplits.ts";
 
-import { assertEquals } from "jsr:@std/assert@~1.0.3/equals";
+import { assertEquals } from "@std/assert/equals";
 
 Deno.test({
   name: "split",
